pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "opendpp-conformance-kit"
version = "0.1.0"
description = "Industrial-grade conformance kit for Digital Product Passports (DPP) and AAS."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "OpenDPP Contributors"}
]
dependencies = [
    "rdflib>=7.0.0",
    "pydash>=8.0.0",
    "pyshacl>=0.25.0",
    "pydantic>=2.5.0",
    "pyld>=2.0.4",
    "jsonschema>=4.18.0",
    "requests>=2.31.0",
    "joserfc>=0.9.0",
    "cryptography>=41.0.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "pyyaml>=6.0.0",
    "jsonpath-ng>=1.6.0",
    "aas-core3.0>=1.0.0",
    "pyld>=2.0.0",
    "jsonschema>=4.17.0",
]

[project.scripts]
dppctl = "opendpp.cli:cli"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
